+++ src/tiled/stampbrush.cpp
@@ -77,7 +77,8 @@ void StampBrush::activate(MapScene *scene)
 
 void StampBrush::deactivate(MapScene *scene)
 {
-    mBrushBehavior = Free;
+    mBrushBehavior = BrushBehavior::Neutral;
+    mBrushState = BrushState::Free;
     mCaptureStampHelper.reset();
     mStampActions->setEnabled(false);
     AbstractTileTool::deactivate(scene);
@@ -85,7 +86,7 @@ void StampBrush::deactivate(MapScene *scene)
 
 void StampBrush::tilePositionChanged(QPoint pos)
 {
-    if (mBrushBehavior == Paint) {
+    if (mBrushState == BrushState::Paint) {
         // Draw a line from the previous point to avoid gaps, skipping the
         // first point, since it was painted when the mouse was pressed, or the
         // last time the mouse was moved.
@@ -117,33 +118,33 @@ void StampBrush::mousePressed(QGraphicsSceneMouseEvent *event)
 {
     if (brushItem()->isVisible()) {
         if (event->button() == Qt::LeftButton) {
-            switch (mBrushBehavior) {
-            case Line:
-                mStampReference = tilePosition();
-                mBrushBehavior = LineStartSet;
-                break;
-            case Circle:
-                mStampReference = tilePosition();
-                mBrushBehavior = CircleMidSet;
-                break;
-            case LineStartSet:
+            switch (mBrushState) {
+            case BrushState::StartSet:
                 doPaint();
                 mStampReference = tilePosition();
                 break;
-            case CircleMidSet:
-                doPaint();
-                break;
-            case Paint:
+            case BrushState::Paint:
                 beginPaint();
                 break;
-            case Free:
-                beginPaint();
+            case BrushState::Free:
+                switch (mBrushBehavior) {
+                case BrushBehavior::Neutral:
+                    beginPaint();
+                    break;
+                case BrushBehavior::Line:
+                case BrushBehavior::Circle:
+                    mStampReference = tilePosition();
+                    mBrushState = BrushState::StartSet;
+                    break;
+                }
                 break;
-            case Capture:
+            case BrushState::Capture:
                 break;
             }
             return;
-        } else if (event->button() == Qt::RightButton && event->modifiers() == Qt::NoModifier) {
+        } else if (event->button() == Qt::RightButton &&
+                   !(event->modifiers() & Qt::ControlModifier))
+        {
             beginCapture();
             return;
         }
@@ -154,32 +155,24 @@ void StampBrush::mousePressed(QGraphicsSceneMouseEvent *event)
 
 void StampBrush::mouseReleased(QGraphicsSceneMouseEvent *event)
 {
-    switch (mBrushBehavior) {
-    case LineStartSet:
+    switch (mBrushState) {
+    case BrushState::StartSet:
         if (event->button() == Qt::LeftButton) {
             if (mStampReference != tilePosition()) {
                 doPaint();
-                mBrushBehavior = Line;
+                mBrushState = BrushState::Free;
             }
         }
         break;
-    case CircleMidSet:
-        if (event->button() == Qt::LeftButton) {
-            if (mStampReference != tilePosition()) {
-                doPaint();
-                updateBrushBehavior();
-            }
-        }
-        break;
-    case Capture:
+    case BrushState::Capture:
         if (event->button() == Qt::RightButton) {
             endCapture();
-            mBrushBehavior = Free;
+            mBrushState = BrushState::Free;
         }
         break;
-    case Paint:
+    case BrushState::Paint:
         if (event->button() == Qt::LeftButton) {
-            mBrushBehavior = Free;
+            mBrushState = BrushState::Free;
 
             // allow going over different variations by repeatedly clicking
             updatePreview();
@@ -194,35 +187,28 @@ void StampBrush::mouseReleased(QGraphicsSceneMouseEvent *event)
 void StampBrush::modifiersChanged(Qt::KeyboardModifiers modifiers)
 {
     mModifiers = modifiers;
-
-    if (!mStamp.isEmpty() || mIsWangFill)
-        updateBrushBehavior();
+    updateBrushBehavior();
 }
 
 void StampBrush::updateBrushBehavior()
 {
     BrushBehavior brushBehavior = mBrushBehavior;
+    BrushState brushState = mBrushState;
 
     if (mModifiers & Qt::ShiftModifier) {
-        if (mModifiers & Qt::ControlModifier) {
-            if (brushBehavior == LineStartSet) {
-                brushBehavior = CircleMidSet;
-            } else if (brushBehavior != CircleMidSet) {
-                brushBehavior = Circle;
-            }
-        } else {
-            if (brushBehavior == CircleMidSet) {
-                brushBehavior = LineStartSet;
-            } else if (brushBehavior != LineStartSet) {
-                brushBehavior = Line;
-            }
-        }
-    } else if (brushBehavior != Paint && brushBehavior != Capture) {
-        brushBehavior = Free;
+        if (mModifiers & Qt::ControlModifier)
+            brushBehavior = BrushBehavior::Circle;
+        else
+            brushBehavior = BrushBehavior::Line;
+    } else {
+        brushBehavior = BrushBehavior::Neutral;
+        if (brushState == BrushState::StartSet)
+            brushState = BrushState::Free;
     }
 
-    if (mBrushBehavior != brushBehavior) {
+    if (brushBehavior != mBrushBehavior || brushState != mBrushState) {
         mBrushBehavior = brushBehavior;
+        mBrushState = brushState;
         updatePreview();
     }
 }
@@ -316,19 +302,19 @@ void StampBrush::setWangSet(WangSet *wangSet)
 
 void StampBrush::beginPaint()
 {
-    if (mBrushBehavior != Free)
+    if (mBrushState != BrushState::Free)
         return;
 
-    mBrushBehavior = Paint;
+    mBrushState = BrushState::Paint;
     doPaint();
 }
 
 void StampBrush::beginCapture()
 {
-    if (mBrushBehavior != Free)
+    if (mBrushState != BrushState::Free)
         return;
 
-    mBrushBehavior = Capture;
+    mBrushState = BrushState::Capture;
     mCaptureStampHelper.beginCapture(tilePosition());
 
     setStamp(TileStamp());
@@ -336,14 +322,15 @@ void StampBrush::beginCapture()
 
 void StampBrush::endCapture()
 {
-    if (mBrushBehavior != Capture)
+    if (mBrushState != BrushState::Capture)
         return;
 
-    mBrushBehavior = Free;
+    mBrushState = BrushState::Free;
 
-    TileStamp stamp = mCaptureStampHelper.endCapture(*mapDocument(), tilePosition());
+    const bool cut = mModifiers & Qt::ShiftModifier;
+    TileStamp stamp = mCaptureStampHelper.endCapture(*mapDocument(), tilePosition(), cut);
     if (!stamp.isEmpty())
-        emit stampChanged(TileStamp(stamp));
+        emit stampChanged(stamp);
     else
         updatePreview();
 }
@@ -609,31 +596,32 @@ void StampBrush::updatePreview(QPoint tilePos)
 
     QRegion tileRegion;
 
-    if (mBrushBehavior == Capture) {
+    if (mBrushState == BrushState::Capture) {
         mPreviewMap.clear();
         tileRegion = mCaptureStampHelper.capturedArea(tilePos);
     } else {
-        switch (mBrushBehavior) {
-        case LineStartSet:
-            drawPreviewLayer(pointsOnLine(mStampReference, tilePos));
+        switch (mBrushState) {
+        case BrushState::StartSet:
+            if (mBrushBehavior == BrushBehavior::Circle) {
+                drawPreviewLayer(pointsOnEllipse(mStampReference,
+                                                 qAbs(mStampReference.x() - tilePos.x()),
+                                                 qAbs(mStampReference.y() - tilePos.y())));
+            } else {
+                drawPreviewLayer(pointsOnLine(mStampReference, tilePos));
+            }
             break;
-        case CircleMidSet:
-            drawPreviewLayer(pointsOnEllipse(mStampReference,
-                                             qAbs(mStampReference.x() - tilePos.x()),
-                                             qAbs(mStampReference.y() - tilePos.y())));
-            break;
-        case Capture:
+        case BrushState::Capture:
             // already handled above
             break;
-        case Circle:
+        case BrushState::Free:
+        case BrushState::Paint:
             // while finding the mid point, there is no need to show
             // the (maybe bigger than 1x1) stamp
-            mPreviewMap.clear();
-            break;
-        case Line:
-        case Free:
-        case Paint:
-            drawPreviewLayer(QVector<QPoint>() << tilePos);
+            if (mBrushBehavior == BrushBehavior::Circle) {
+                mPreviewMap.clear();
+            } else {
+                drawPreviewLayer(QVector<QPoint>() << tilePos);
+            }
             break;
         }